test(audit): give connection.created its own live-stack suite
The three creation cases moved out of the general audit suite, which is
about the auditable() middleware contract - wiring order, resolve-before-
next, target-from-response - none of which applies to an event emitted from
a hook. They now sit beside the module and unit file they cover, and the
general suite drops from 680 to 561 lines.

The remaining cases in that block genuinely are targetFromResponse cases on
the middleware, so they stayed.
